# go/api_default.go
/*
 * Provisional Term Management API
 *
 * DRAFT API specification for provisional term request management for an ontology. An instance of the API is attached to a particular ontology, and configured on the back-end to interface with the tooling surrounding that ontology. Therefore, information does not have to be given through the API as to location of Ontology/GitHub/Wiki instances for term addition.
 *
 * API version: 1.0.0
 * Contact: l.slater.1@bham.ac.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"net/http"
)

func RequestTermAddition(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
}

func RequestTermInclusion(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
}