# go/model_new_term_request.go
/*
 * Provisional Term Management API
 *
 * DRAFT API specification for provisional term request management for an ontology. An instance of the API is attached to a particular ontology, and configured on the back-end to interface with the tooling surrounding that ontology. Therefore, information does not have to be given through the API as to location of Ontology/GitHub/Wiki instances for term addition.
 *
 * API version: 1.0.0
 * Contact: l.slater.1@bham.ac.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type NewTermRequest struct {

	// The first or primary label to be associated with the new class.
	FirstLabel string `json:"firstLabel"`

	// Any additional labels or synonyms to be associated with the new class.
	Synonyms []string `json:"synonyms,omitempty"`

	// A textual definition or description of the new class.
	Description string `json:"description"`

	// Name of the person requesting the term addition.
	CreatorName string `json:"creatorName,omitempty"`

	// Email address of the person requesting the term addition.
	CreatorEmail string `json:"creatorEmail,omitempty"`

	// The IRI of the suggested direct parent class for the requested term.
	ParentClass string `json:"parentClass"`
}